rag-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0
qdrant-client>=1.7.0
llama-index>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-vector-stores-qdrant>=0.1.0
sentence-transformers>=2.2.0
psycopg2-binary
sqlalchemy>=2.0.0
alembic>=1.10.0
asyncpg>=0.25.0
